refactor: reorganize imports and update control table references for Dynamixel motors
