+++ src/robopy/robots/rakuda/rakuda_pair_sys.py
@@ -11,8 +11,8 @@ from robopy.config.robot_config.rakuda_config import (
     RakudaArmObs,
     RakudaConfig,
 )
-from robopy.motor.control_table import XControlTable
 from robopy.motor.dynamixel_bus import DynamixelBus
+from robopy.motor.dynamixel_control_table import XControlTable
 
 from ..common.robot import Robot
 from .rakuda_follower import RakudaFollower
